requirements.txt
python-dotenv
numpy
pandas
seaborn
matplotlib
tqdm
librosa
ipython
ipykernel
scikit-learn
dataclasses
ensure
python-box
PyYAML
-e .